Discovered a few new resources online, now attempting a fairly radical remodel of the system

+++ JawJaw/src/pitchDetection/Mode.java
@@ -2,8 +2,8 @@ package pitchDetection;
 
 public class Mode {
 	
-	 public float getMode(float[] n){//method for calculating the mode of the pitch array generated
-	    float t = 0;
+	 public double getMode(double[] n){//method for calculating the mode of the pitch array generated
+	    double t = 0;
 	    for(int i=0; i<n.length; i++){
 	        for(int j=1; j<n.length-i; j++){
 	            if(n[j-1] > n[j]){
@@ -14,7 +14,7 @@ public class Mode {
 	        }
 	    }
 
-	    float mode = n[0];
+	    double mode = n[0];
 	    int temp = 1;
 	    int temp2 = 1;
 	    for(int i=1;i<n.length;i++){
